fastapi/requirements.txt
fastapi
pillow
torch
torchvision
python-multipart
aiofiles
pytorch-lightning
lightning-bolts
captum
Pillow